# optimization: join probe returns first item directly