--- a/hydroflow/src/compiled/pull/half_join_state/set.rs
+++ b/hydroflow/src/compiled/pull/half_join_state/set.rs
@@ -62,17 +62,21 @@ where
         false
     }
 
-    fn probe(&mut self, k: &Key, v: &ValProbe) {
-        if let Some(entry) = self.table.get(k) {
-            // TODO: We currently don't free/shrink the self.current_matches vecdeque to save time.
-            // This mean it will grow to eventually become the largest number of matches in a single probe call.
-            // Maybe we should clear this memory at the beginning of every tick/periodically?
-            self.current_matches.extend(
-                entry
-                    .iter()
-                    .map(|valbuild| (k.clone(), v.clone(), valbuild.clone())),
-            );
-        }
+    fn probe(&mut self, k: &Key, v: &ValProbe) -> Option<(Key, ValProbe, ValBuild)> {
+        // TODO: We currently don't free/shrink the self.current_matches vecdeque to save time.
+        // This mean it will grow to eventually become the largest number of matches in a single probe call.
+        // Maybe we should clear this memory at the beginning of every tick/periodically?'
+        let mut iter = self
+            .table
+            .get(k)?
+            .iter()
+            .map(|valbuild| (k.clone(), v.clone(), valbuild.clone()));
+
+        let first = iter.next();
+
+        self.current_matches.extend(iter);
+
+        first
     }
 
     fn pop_match(&mut self) -> Option<(Key, ValProbe, ValBuild)> {