evidence for this change is pretty weak:

```
micro/ops/join          time:   [113.65 µs 114.38 µs 115.80 µs]
                        change: [-0.0148% +1.0042% +2.3994%] (p = 0.12 > 0.05)
                        No change in performance detected.

micro/ops/crossjoin     time:   [152.13 µs 153.46 µs 155.27 µs]
                        change: [+2.3725% +3.6797% +5.3980%] (p = 0.00 < 0.05)
                        Performance has regressed.

micro/ops/anti_join     time:   [12.927 µs 12.957 µs 12.994 µs]
                        change: [-0.4503% +0.1853% +0.7231%] (p = 0.57 > 0.05)
                        No change in performance detected.

symmetric_hash_join/no_match
                        time:   [126.19 µs 126.68 µs 127.37 µs]
                        change: [-4.5148% -0.9976% +2.6660%] (p = 0.59 > 0.05)
                        No change in performance detected.

symmetric_hash_join/match_keys_diff_values
                        time:   [174.17 µs 174.47 µs 174.78 µs]
                        change: [-6.0184% -3.7568% -1.7607%] (p = 0.00 < 0.05)
                        Performance has improved.

symmetric_hash_join/match_keys_same_values
                        time:   [173.95 µs 174.89 µs 176.70 µs]
                        change: [-6.2952% -3.9995% -1.7759%] (p = 0.00 < 0.05)
                        Performance has improved.

symmetric_hash_join/zipf_keys_low_contention_unique_values
                        time:   [164.71 µs 166.60 µs 168.60 µs]
                        change: [-4.2218% -1.9710% +0.2797%] (p = 0.10 > 0.05)
                        No change in performance detected.

symmetric_hash_join/zipf_keys_high_contention_unique_values
                        time:   [3.4813 ms 3.5060 ms 3.5381 ms]
                        change: [-2.8375% -1.1905% +0.3146%] (p = 0.14 > 0.05)
                        No change in performance detected.
```